Creacion Diplomas realizado
Todos los try catch ya creados, mas creacion de diplomas y el subir archivos, cambios en el sp_CraerDiploma, devuelve una lista

+++ MinisteriodeEducacionMVCApp/MinisteriodeEducacionMVCApp/Controllers/ListadeEstudiantesController.cs
@@ -145,10 +145,19 @@ namespace MinisteriodeEducacionMVCApp.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            ListadeEstudiantes listadeEstudiantes = db.ListadeEstudiantes.Find(id);
-            db.ListadeEstudiantes.Remove(listadeEstudiantes);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            var Del = db.Estudiante.Where(x => x.idListaEstudiante == id).Count();
+            if (Del == 1)
+            {
+                TempData["DErr2"] = "Estudiante con Login en la APP: Borrar primero Cuenta de Estudiante";
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                ListadeEstudiantes listadeEstudiantes = db.ListadeEstudiantes.Find(id);
+                db.ListadeEstudiantes.Remove(listadeEstudiantes);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
         }
         [HttpGet]
         public ActionResult ImportarLista()
